Turn all files in src/ into their own modules and export functions
It is a little less than ideal when some modules depend on each other, and could possibly be refactored so the structs are defined in their repective modules, but having everything as a module makes it simpler to avoid polluting the namespace with constants that may conflict.
Importing everything from a single module that depends on submodules makes it easier to load the file in a way that allows the language server to recognize references (cannot use joinpath and must specify absolute or relative path, which will be different in the REPL vs in the filesystem).