Start trying to use Makefiles to run pipeline
Trying to figure out how I can use automatic variables to run multiple files with the same pattern.

Seems to require a bit of a hack (creating a tmp file) to be able to run a script without explicitly using make to produce the output.
This is necessary as I want to be able to continue using @produce_or_load to compute multiple simulations by parameter combinations, rather than starting a new Julia session for each one and use Make to save the output