Rewrite detection delay function to produce arrays of all pairs
Each outbreak may receive multiple alerts, so want to return all the matches, not just the first.
The delay is calculated by the first alert, and can be negative.
This updated formula allows for calculation of the number of false positives as well as the number of alerts per outbreak